+++ elrond-wasm-debug/src/tx_execution/builtin_function_mocks/esdt_nft/esdt_nft_add_uri_mock.rs
@@ -5,7 +5,7 @@ use elrond_wasm::{
 
 use crate::tx_mock::{BlockchainUpdate, TxCache, TxInput, TxLog, TxResult, TxResultCalls};
 
-use super::builtin_func_trait::BuiltinFunction;
+use super::super::builtin_func_trait::BuiltinFunction;
 
 pub struct ESDTNftAddUri;
