+++ elrond-wasm-debug/src/tx_execution/builtin_function_mocks/esdt_nft/esdt_nft_create_mock.rs
@@ -9,7 +9,7 @@ use crate::{
     world_mock::{EsdtInstance, EsdtInstanceMetadata},
 };
 
-use super::builtin_func_trait::BuiltinFunction;
+use super::super::builtin_func_trait::BuiltinFunction;
 
 pub struct ESDTNftCreate;
