+++ elrond-wasm-debug/src/tx_execution/builtin_function_mocks/esdt_nft/mod.rs
@@ -1,0 +1,15 @@
+mod esdt_local_burn;
+mod esdt_local_mint;
+mod esdt_nft_add_quantity_mock;
+mod esdt_nft_add_uri_mock;
+mod esdt_nft_burn_mock;
+mod esdt_nft_create_mock;
+mod esdt_nft_update_attriutes_mock;
+
+pub use esdt_local_burn::*;
+pub use esdt_local_mint::*;
+pub use esdt_nft_add_quantity_mock::*;
+pub use esdt_nft_add_uri_mock::*;
+pub use esdt_nft_burn_mock::*;
+pub use esdt_nft_create_mock::*;
+pub use esdt_nft_update_attriutes_mock::*;
