+++ elrond-wasm-debug/src/tx_execution/builtin_function_mocks/general/claim_developer_rewards_mock.rs
@@ -4,7 +4,7 @@ use num_traits::Zero;
 
 use crate::tx_mock::{BlockchainUpdate, TxCache, TxInput, TxResult};
 
-use super::builtin_func_trait::BuiltinFunction;
+use super::super::builtin_func_trait::BuiltinFunction;
 
 pub struct ClaimDeveloperRewards;
