+++ elrond-wasm-debug/src/tx_execution/builtin_function_mocks/general/mod.rs
@@ -1,0 +1,9 @@
+mod change_owner_mock;
+mod claim_developer_rewards_mock;
+mod set_username_mock;
+mod upgrade_contract;
+
+pub use change_owner_mock::*;
+pub use claim_developer_rewards_mock::*;
+pub use set_username_mock::*;
+pub use upgrade_contract::*;
