+++ elrond-wasm-debug/src/tx_execution/builtin_function_mocks/mod.rs
@@ -2,20 +2,9 @@ mod builtin_func_exec;
 mod builtin_func_map;
 mod builtin_func_role_check_wrapper;
 mod builtin_func_trait;
-mod change_owner_mock;
-mod claim_developer_rewards_mock;
-mod esdt_local_burn;
-mod esdt_local_mint;
-mod esdt_multi_transfer_mock;
-mod esdt_nft_add_quantity_mock;
-mod esdt_nft_add_uri_mock;
-mod esdt_nft_burn_mock;
-mod esdt_nft_create_mock;
-mod esdt_nft_transfer_mock;
-mod esdt_nft_update_attriutes_mock;
-mod esdt_transfer_mock;
-mod set_username_mock;
-mod upgrade_contract;
+mod esdt_nft;
+mod general;
+mod transfer;
 
 pub use builtin_func_exec::{execute_builtin_function_or_default, init_builtin_functions};
 pub use builtin_func_map::BuiltinFunctionMap;
