+++ elrond-wasm-debug/src/tx_execution/builtin_function_mocks/transfer/esdt_multi_transfer_mock.rs
@@ -9,7 +9,7 @@ use crate::{
     tx_mock::{BlockchainUpdate, TxCache, TxInput, TxInputESDT, TxLog, TxResult},
 };
 
-use super::builtin_func_trait::BuiltinFunction;
+use super::super::builtin_func_trait::BuiltinFunction;
 
 pub struct ESDTMultiTransfer;
