+++ elrond-wasm-debug/src/tx_execution/builtin_function_mocks/transfer/mod.rs
@@ -1,0 +1,7 @@
+mod esdt_multi_transfer_mock;
+mod esdt_nft_transfer_mock;
+mod esdt_transfer_mock;
+
+pub use esdt_multi_transfer_mock::*;
+pub use esdt_nft_transfer_mock::*;
+pub use esdt_transfer_mock::*;
